fullscreen layout

+++ skycapitalsservers/src/main/java/com/nikego/skycapitals/controllers/CardController.java
@@ -2,10 +2,15 @@ package com.nikego.skycapitals.controllers;
 
 import com.nikego.skycapitals.models.Card;
 import com.nikego.skycapitals.services.CardService;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.RestController;
 
 import java.util.AbstractMap;
 
@@ -21,7 +26,7 @@ public class CardController {
 
     @GetMapping(value="/readCard/{id}",consumes = "application/json",produces = "application/json")
     @ResponseBody
-    public ResponseEntity<?> read(@PathVariable(name = "id") int id){
+    public ResponseEntity<?> read(@PathVariable(name = "id") Long id){
         final Card card=cardService.read(id);
         return card != null
                 ? ResponseEntity.status(HttpStatus.OK).body(new AbstractMap.SimpleEntry<>("Карта:",card))
