+++ skycapitalsservers/src/main/java/com/nikego/skycapitals/repository/CardRepository.java
@@ -1,9 +1,10 @@
 package com.nikego.skycapitals.repository;
 
 import com.nikego.skycapitals.models.Card;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CardRepository extends JpaRepository<Card,Integer> {
+public interface CardRepository extends JpaRepository<Card,Long> {
     boolean existsByNameCard(String nameCard);
     boolean existsByNumberCard(long numberCard);
 }
