+++ skycapitalsservers/src/main/java/com/nikego/skycapitals/services/CardService.java
@@ -6,7 +6,7 @@ import java.util.Optional;
 
 public interface CardService {
 
-    Card read(int id);
+    Card read(Long id);
 
     boolean update(Card card, long numberCard);
