+++ skycapitalsservers/src/main/java/com/nikego/skycapitals/services/IUserTransaction.java
@@ -1,10 +1,8 @@
 package com.nikego.skycapitals.services;
 
-import com.nikego.skycapitals.models.Card;
-
 public interface IUserTransaction {
 
     /*перевод с карты на карту*/
-    boolean transferMoney(Integer numberCardSender, Integer numberCardRecipient, int sum);
+    boolean transferMoney(Long numberCardSender, Long numberCardRecipient, int sum);
 
 }
